Creates a struct style for button to apply themes
