+++ Sources/Public/Components/Button/NatButton/NatButton+Style.swift
@@ -1,0 +1,17 @@
+extension NatButton {
+    public struct Style {
+        let apply: (NatButton) -> Void
+        let changeState: (NatButton) -> Void
+        let applyTitle: (String?, NatButton) -> Void
+    }
+}
+
+extension NatButton.Style {
+    public static var outlined: NatButton.Style {
+        .init(
+            apply: ButtonOutlinedStyle.applyStyle,
+            changeState: ButtonOutlinedStyle.applyStyeForStateChanges,
+            applyTitle: ButtonOutlinedStyle.applyStyleForTitle
+        )
+    }
+}
